+++ pages/mycollections.tsx
@@ -1,297 +1,44 @@
-import type { FC, MouseEvent } from 'react'
+import type { FC } from 'react'
 import type Collection from '../interfaces/Collection'
 
 import React, { useState, useContext, useEffect } from 'react'
-import { useRouter } from 'next/navigation'
 
 import AppContext from '../context/AppContext'
 
 import Header from '../app/components/NavBar/Header'
 import CreatorHeader from '../app/components/Creator/CreatorHeader'
-import CollectionModal from '../app/components/collection/CollectionModal'
+import CollectionModal from '../app/components/CollectionModal/CollectionModal'
+import CreateCollectionButton from '../app/components/MyCollections/CreateCollectionButton'
+import CollectionItem from '../app/components/MyCollections/CollectionItem'
 
 import useGetCollections from '../hooks/useGetCollections'
+import useGetProfile from '../hooks/useGetProfile'
 
 import styles from '../styles/MyCollections.module.css'
 
-import { LoadingOutlined, PlusOutlined } from '@ant-design/icons'
-import { AiTwotoneEdit } from 'react-icons/ai'
-import { TiDelete } from 'react-icons/ti'
-import { MdMoreVert } from 'react-icons/md'
-import type { MenuProps } from 'antd'
-import {
-	Typography,
-	Button,
-	Row,
-	Spin,
-	Col,
-	Dropdown,
-	ConfigProvider,
-	theme,
-} from 'antd'
-
-const themeDark = { algorithm: theme.darkAlgorithm }
-const themeDefault = { algorithm: theme.defaultAlgorithm }
+import { LoadingOutlined } from '@ant-design/icons'
+import { Row, Spin, Col } from 'antd'
 
 const antIcon = <LoadingOutlined style={{ fontSize: 24 }} spin />
-const { Text } = Typography
-
-const CreateCollectionButton: FC = ({ currentTheme }) => {
-	const context = useContext(AppContext)
-
-	const setIsCollectionModalOpen =
-		context?.setIsCollectionModalOpen ?? (() => {})
-	const setCollectionModalView = context?.setCollectionModalView ?? (() => {})
-
-	const handleCreateCollection = (): void => {
-		setCollectionModalView('create')
-		setIsCollectionModalOpen(true)
-	}
-
-	return (
-		<Row className={styles.createCollectionButtonWrapper}>
-			<Button
-				className={styles.createCollectionButton}
-				type='default'
-				shape='round'
-				size='large'
-				icon={<PlusOutlined />}
-				onClick={(e) => {
-					handleCreateCollection()
-				}}
-			>
-				<Text className={styles.createCollectionButtonText}>
-					{'Create Collection'}
-				</Text>
-			</Button>
-		</Row>
-	)
-}
-
-interface CollectionItemButton {
-	isCollectionHovering: boolean
-	collection: Collection
-}
-
-const CollectionItemMoreButton: FC<CollectionItemButton> = ({
-	isCollectionHovering,
-	collection,
-}) => {
-	const handleMoreCollectionModal = (): void => {
-		console.log('handleMoreCollection')
-	}
-
-	const items: MenuProps['items'] = [
-		{
-			key: 1,
-			label: (
-				<Text onClick={() => handleDeleteCollectionModal}>{'Delete'}</Text>
-			),
-		},
-	]
-
-	return (
-		<>
-			{isCollectionHovering ? (
-				<span>
-					<Dropdown menu={{ items }}>
-						<Button
-							// onClick={handleButtonClick}
-							shape='circle'
-							size='large'
-							type='text'
-							icon={<MdMoreVert />}
-						/>
-					</Dropdown>
-				</span>
-			) : null}
-		</>
-	)
-}
-
-const CollectionItemDeleteButton: FC<CollectionItemButton> = ({
-	collection,
-	currentTheme,
-}) => {
-	const context = useContext(AppContext)
-
-	const setIsCollectionModalOpen =
-		context?.setIsCollectionModalOpen ?? (() => {})
-	const setCollectionModalView = context?.setCollectionModalView ?? (() => {})
-	const setCurrentModalCollection =
-		context?.setCurrentModalCollection ?? (() => {})
-
-	const handleDeleteCollectionModal = (): void => {
-		console.log('handleDeleteCollection')
-		console.log({ collection })
-		console.log(collection.name)
-		setCollectionModalView('delete')
-		setIsCollectionModalOpen(true)
-		setCurrentModalCollection(collection)
-	}
-
-	const deleteStyles = {
-		display: 'flex',
-		alignItems: 'center',
-		justifyContent: 'center ',
-		marginLeft: 10,
-	}
-
-	return (
-		<>
-			<span>
-				<Button
-					onClick={handleDeleteCollectionModal}
-					shape='circle'
-					size='large'
-					type='text'
-					style={deleteStyles}
-					icon={<TiDelete />}
-				/>
-			</span>
-		</>
-	)
-}
-
-const CollectionItemRenameButton: FC<CollectionItemButton> = ({
-	collection,
-	currentTheme,
-}) => {
-	const renameStyles = {
-		display: 'flex',
-		alignItems: 'center',
-		justifyContent: 'center ',
-		marginLeft: 10,
-	}
-
-	const context = useContext(AppContext)
-
-	const setIsCollectionModalOpen =
-		context?.setIsCollectionModalOpen ?? (() => {})
-	const setCollectionModalView = context?.setCollectionModalView ?? (() => {})
-	const setCurrentModalCollection =
-		context?.setCurrentModalCollection ?? (() => {})
-
-	const handleRenameCollectionModal = (): void => {
-		console.log('handleDeleteCollection')
-		console.log({ collection })
-		console.log(collection.name)
-		setCollectionModalView('rename')
-		setIsCollectionModalOpen(true)
-		setCurrentModalCollection(collection)
-	}
-
-	return (
-		<>
-			<span>
-				<Button
-					shape='circle'
-					size='large'
-					type='text'
-					icon={<AiTwotoneEdit />}
-					style={renameStyles}
-					onClick={handleRenameCollectionModal}
-				/>
-			</span>
-		</>
-	)
-}
-
-const CollectionItem: FC<Collection> = ({ collection, currentTheme }) => {
-	const router = useRouter()
-
-	const [isCollectionHovering, setIsCollectionHovering] =
-		useState<boolean>(false)
-
-	const handleMouseOver = (): void => {
-		setIsCollectionHovering(true)
-	}
-
-	const handleMouseOut = (): void => {
-		setIsCollectionHovering(false)
-	}
-
-	const handleClickCollection = (
-		e: MouseEvent<HTMLAnchorElement>,
-		collectionId: string
-	): void => {
-		e.preventDefault()
-		router.push(`/collection/${String(collectionId)}`)
-	}
-
-	// console.log({ isCollectionHovering })
-
-	const buttonTheme = currentTheme === 'light' ? themeDefault : themeDark
-
-	return (
-		<section
-			key={collection._id}
-			onMouseOver={handleMouseOver}
-			onMouseOut={handleMouseOut}
-			style={{ position: 'relative' }}
-		>
-			<ConfigProvider theme={buttonTheme}>
-				<Button
-					className={styles.collectionButton}
-					onClick={(e) => {
-						handleClickCollection(e, collection._id)
-					}}
-				>
-					<Text className={styles.collectionButtonText}>{collection.name}</Text>
-				</Button>
-			</ConfigProvider>
-
-			{isCollectionHovering ? (
-				<>
-					<Row
-						style={{
-							position: 'absolute',
-							top: 20,
-							right: 20,
-							zIndex: 100,
-						}}
-					>
-						<CollectionItemRenameButton
-							collection={collection}
-							isCollectionHovering={isCollectionHovering}
-						/>
-					</Row>
-					<Row
-						style={{
-							position: 'absolute',
-							bottom: 20,
-							right: 20,
-							zIndex: 100,
-						}}
-					>
-						<CollectionItemDeleteButton
-							collection={collection}
-							isCollectionHovering={isCollectionHovering}
-						/>
-					</Row>
-				</>
-			) : null}
-		</section>
-	)
-}
 
 const MyCollections: FC = () => {
 	const context = useContext(AppContext)
 
 	const [myCollections, setMyCollections] = useState<Collections>([])
 
-	const userId =
+	const isContext =
 		typeof context !== 'undefined' && context !== null && 'userId' in context
-			? String(context.userId)
-			: ''
+	const userId = isContext ? String(context.userId) : ''
+	const userAddress = context?.userAddress ?? ''
 
 	const currentTheme = context?.currentTheme ?? ''
 
 	const collections = context?.collections ?? []
 	const setCollections = context?.setCollections ?? (() => {})
 
-	const collectionModalView = context?.collectionModalView ?? ''
 	const currentModalCollection = context?.currentModalCollection ?? ''
+
+	const creator = useGetProfile(userId)
 
 	const {
 		collections: myCollectionsData,
@@ -303,26 +50,34 @@ const MyCollections: FC = () => {
 		setMyCollections(collections)
 	}, [collections, setCollections])
 
-	const handleCurrentCollectionsData = (): void => {
+	useEffect(() => {
+		if (myCollectionsData?.collections?.length > 0) {
+			setMyCollections(myCollectionsData?.collections)
+		}
+	}, [myCollectionsData])
+
+	const handleCurrentCollectionsData = (): Collection[] => {
 		if (myCollections.length > 0) {
 			return myCollections
 		} else if (myCollectionsData.length > 0) {
 			return myCollectionsData
 		}
+		return []
 	}
-
-	const isMyCollectionsData =
-		myCollectionsData !== null && !isLoading && typeof error === 'undefined'
-
 	const currentMyCollections = handleCurrentCollectionsData()
-	const isCurrentMyCollections = currentMyCollections.length > 0
+
+	const isCurrentMyCollectionsData =
+		currentMyCollections !== null && !isLoading && typeof error === 'undefined'
+
+	const isCurrentMyCollections = currentMyCollections?.length > 0 ?? []
 
 	console.log({ myCollectionsData, isLoading, error })
-	console.log({ myCollections })
-	console.log({ currentMyCollections })
+	console.log({ creator })
+	// console.log({ myCollections })
+	// console.log({ currentMyCollections })
 
-	console.log({ collectionModalView })
-	console.log({ currentModalCollection })
+	// console.log({ collectionModalView })
+	// console.log({ currentModalCollection })
 
 	const collectionWrapperStyles = {
 		display: 'flex',
@@ -335,6 +90,8 @@ const MyCollections: FC = () => {
 		flexDirection: 'column',
 	}
 
+	console.log({ userAddress })
+
 	return (
 		<>
 			<main>
@@ -342,7 +99,11 @@ const MyCollections: FC = () => {
 			</main>
 
 			<Col style={collectionWrapperStyles}>
-				<CreatorHeader userId={userId} isMyCollectionsRoute={true} />
+				<CreatorHeader
+					creator={creator}
+					userAddress={userAddress}
+					isMyCollectionsRoute={true}
+				/>
 				{isCurrentMyCollections ? (
 					<Col className={styles.collectionsWrapper}>
 						<Row style={collectionInnerWrapperStyles}>
@@ -351,15 +112,20 @@ const MyCollections: FC = () => {
 							<CollectionModal collection={currentModalCollection} />
 
 							<Row style={{ justifyContent: 'center' }}>
-								{currentMyCollections.map((collection: Collection) => {
-									return (
-										<CollectionItem
-											key={collection._id}
-											collection={collection}
-											currentTheme={currentTheme}
-										/>
-									)
-								})}
+								{currentMyCollections?.map(
+									(collection: Collection, index: number) => {
+										return (
+											<CollectionItem
+												key={collection._id}
+												collection={collection}
+												currentTheme={currentTheme}
+												collectionCreations={
+													myCollectionsData?.collectionsCreations[index]
+												}
+											/>
+										)
+									}
+								)}
 							</Row>
 						</Row>
 					</Col>
