+++ styles/MyCollections.module.css
@@ -65,11 +65,17 @@
 	min-height: 200px;
 	border-radius: 10px;
 	margin: 10px;
+	overflow: hidden;
+	box-shadow: 0px 0px 2px rgba(0, 0, 0, 0.15), 0px 2px 5px rgba(0, 0, 0, 0.05),
+		0px 8px 40px rgba(0, 0, 0, 0.04);
 }
 
 .collectionButtonText {
 	font-weight: bold;
 	font-size: 1.2rem;
+	z-index: 50;
+	position: relative;
+	filter: box-shadow(0px 0px 10px rgba(0, 0, 0, 0.5));
 }
 
 .collectionsWrapper {
